Remove subscription-submitter
