+++ src/controllers/admin/webhook.ts
@@ -5,10 +5,24 @@ import { StatusCodes } from 'http-status-codes';
 import { EventTracker, eventTracker } from '../../services/track/tracker.js';
 import type { INotifyMessage } from '../../types/track.js';
 import { OperationNameEnum } from '../../types/constants.js';
-import { buildSubmitOperation } from '../../services/track/helpers.js';
+import { SubscriptionService } from '../../services/admin/subscription.js';
+import { CustomerService } from '../../services/api/customer.js';
+import { LogToHelper } from '../../middleware/auth/logto-helper.js';
+import { UserService } from '../../services/api/user.js';
+import type { SupportedPlanTypes } from '../../types/admin.js';
+import type { CustomerEntity } from '../../database/entities/customer.entity.js';
+import { buildSubscriptionData } from '../../services/track/helpers.js';
 
 dotenv.config();
 export class WebhookController {
+	private readonly stripe: Stripe;
+
+	constructor() {
+		this.stripe = new Stripe(process.env.STRIPE_SECRET_KEY);
+	}
+
+	public static instance = new WebhookController();
+
 	public async handleWebhook(request: Request, response: Response) {
 		// Signature verification and webhook handling is placed in the same method
 		// cause stripe uses the mthod which validate the signature and provides the event.
@@ -57,9 +71,6 @@ export class WebhookController {
 						),
 						severity: 'info',
 					} satisfies INotifyMessage);
-					await eventTracker.submit(
-						buildSubmitOperation(subscription, OperationNameEnum.SUBSCRIPTION_TRIAL_WILL_END)
-					);
 					break;
 				case 'customer.subscription.deleted':
 					subscription = event.data.object;
@@ -71,9 +82,7 @@ export class WebhookController {
 						),
 						severity: 'info',
 					} satisfies INotifyMessage);
-					await eventTracker.submit(
-						buildSubmitOperation(subscription, OperationNameEnum.SUBSCRIPTION_CANCEL)
-					);
+					await this.handleSubscriptionCancel(subscription);
 					break;
 				case 'customer.subscription.created':
 					subscription = event.data.object;
@@ -85,9 +94,7 @@ export class WebhookController {
 						),
 						severity: 'info',
 					} satisfies INotifyMessage);
-					await eventTracker.submit(
-						buildSubmitOperation(subscription, OperationNameEnum.SUBSCRIPTION_CREATE)
-					);
+					await this.handleSubscriptionCreate(subscription);
 					break;
 				case 'customer.subscription.updated':
 					subscription = event.data.object;
@@ -99,9 +106,7 @@ export class WebhookController {
 						),
 						severity: 'info',
 					} satisfies INotifyMessage);
-					await eventTracker.submit(
-						buildSubmitOperation(subscription, OperationNameEnum.SUBSCRIPTION_UPDATE)
-					);
+					await this.handleSubscriptionUpdate(subscription);
 					break;
 				default:
 					// Unexpected event type
@@ -129,5 +134,292 @@ export class WebhookController {
 				error: `Internal error: ${(error as Error)?.message || error}`,
 			});
 		}
+	}
+
+	async handleSubscriptionCreate(subscription: Stripe.Subscription, customer?: CustomerEntity): Promise<void> {
+		const data = buildSubscriptionData(subscription);
+		const operation = OperationNameEnum.SUBSCRIPTION_CREATE;
+		try {
+			const [product, stripeCustomer] = await Promise.all([
+				this.stripe.products.retrieve(data.productId),
+				this.stripe.customers.retrieve(data.paymentProviderId),
+			]);
+			if (!customer) {
+				const customers = await CustomerService.instance.customerRepository.find({
+					where: { paymentProviderId: data.paymentProviderId },
+				});
+				if (customers.length === 0) {
+					// we add an additional check in case that a customer was created locally with email and no paymentProviderId
+					if (!stripeCustomer.deleted && stripeCustomer.email) {
+						const customerWithoutPaymentProviderId =
+							await CustomerService.instance.customerRepository.findOne({
+								where: { email: stripeCustomer.email },
+							});
+
+						if (!customerWithoutPaymentProviderId) {
+							await eventTracker.notify({
+								message: EventTracker.compileBasicNotification(
+									`Customer not found for Cheqd Studio, creating new customer record with paymentProviderId: ${data.paymentProviderId}`,
+									operation
+								),
+								severity: 'info',
+							});
+
+							const customerName = stripeCustomer.name ?? stripeCustomer.email;
+							customers.push(
+								await CustomerService.instance.create(
+									customerName,
+									stripeCustomer.email,
+									undefined,
+									data.paymentProviderId
+								)
+							);
+						} else {
+							customers.push(customerWithoutPaymentProviderId);
+						}
+					} else {
+						await eventTracker.notify({
+							message: EventTracker.compileBasicNotification(
+								`Customer not found for Cheqd Studio, cannot create new customer without a email id: ${data.paymentProviderId}`,
+								operation
+							),
+							severity: 'error',
+						});
+						return;
+					}
+				} else if (customers.length !== 1) {
+					await eventTracker.notify({
+						message: EventTracker.compileBasicNotification(
+							`Only one Stripe account should be associated with CaaS customer. Stripe accountId: ${data.paymentProviderId}.`,
+							operation
+						),
+						severity: 'error',
+					});
+				}
+
+				customer = customers[0];
+			}
+
+			const subscription = await SubscriptionService.instance.create(
+				data.subscriptionId,
+				customer,
+				data.status,
+				data.currentPeriodStart,
+				data.currentPeriodEnd,
+				data.trialStart as Date,
+				data.trialEnd as Date
+			);
+			if (!subscription) {
+				await eventTracker.notify({
+					message: EventTracker.compileBasicNotification(
+						`Failed to create a new subscription with id: ${data.subscriptionId}.`,
+						operation
+					),
+					severity: 'error',
+				});
+			}
+
+			await this.syncLogtoRoles(operation, customer.customerId, product.name);
+
+			await eventTracker.notify({
+				message: EventTracker.compileBasicNotification(
+					`Subscription created with id: ${data.subscriptionId}.`,
+					operation
+				),
+				severity: 'info',
+			});
+		} catch (error) {
+			await eventTracker.notify({
+				message: EventTracker.compileBasicNotification(
+					`Failed to create a new subscription with id: ${data.subscriptionId} because of error: ${(error as Error)?.message || error}`,
+					operation
+				),
+				severity: 'error',
+			});
+		}
+	}
+
+	async handleSubscriptionUpdate(subscription: Stripe.Subscription): Promise<void> {
+		const data = buildSubscriptionData(subscription);
+		const operation = OperationNameEnum.SUBSCRIPTION_UPDATE;
+		try {
+			const subscription = await SubscriptionService.instance.update(
+				data.subscriptionId,
+				data.status,
+				data.currentPeriodStart,
+				data.currentPeriodEnd,
+				data.trialStart as Date,
+				data.trialEnd as Date
+			);
+			if (!subscription) {
+				await eventTracker.notify({
+					message: EventTracker.compileBasicNotification(
+						`Failed to update subscription with id: ${data.subscriptionId}.`,
+						operation
+					),
+					severity: 'error',
+				});
+			}
+
+			const [customer, product] = await Promise.all([
+				CustomerService.instance.findbyPaymentProviderId(data.paymentProviderId),
+				this.stripe.products.retrieve(data.productId),
+			]);
+
+			if (customer) {
+				await this.syncLogtoRoles(operation, customer.customerId, product.name);
+			}
+
+			await eventTracker.notify({
+				message: EventTracker.compileBasicNotification(
+					`Subscription updated with id: ${data.subscriptionId}.`,
+					operation
+				),
+				severity: 'info',
+			});
+		} catch (error) {
+			await eventTracker.notify({
+				message: EventTracker.compileBasicNotification(
+					`Failed to update subscription with id: ${data.subscriptionId} because of error: ${(error as Error)?.message || error}`,
+					operation
+				),
+				severity: 'error',
+			});
+		}
+	}
+
+	async handleSubscriptionCancel(subscription: Stripe.Subscription): Promise<void> {
+		const data = buildSubscriptionData(subscription);
+		const operation = OperationNameEnum.SUBSCRIPTION_CANCEL;
+		try {
+			const subscription = await SubscriptionService.instance.update(data.subscriptionId, data.status);
+			if (!subscription) {
+				await eventTracker.notify({
+					message: EventTracker.compileBasicNotification(
+						`Failed to cancel subscription with id: ${data.subscriptionId}.`,
+						operation
+					),
+					severity: 'error',
+				});
+			}
+
+			const customer = await CustomerService.instance.findbyPaymentProviderId(data.paymentProviderId);
+			if (customer) {
+				this.syncLogtoRoles(operation, customer.customerId, '');
+			}
+
+			await eventTracker.notify({
+				message: EventTracker.compileBasicNotification(
+					`Subscription canceled with id: ${data.subscriptionId}.`,
+					operation
+				),
+				severity: 'info',
+			});
+		} catch (error) {
+			await eventTracker.notify({
+				message: EventTracker.compileBasicNotification(
+					`Failed to cancel subscription with id: ${data.subscriptionId} because of error: ${(error as Error)?.message || error}`,
+					operation
+				),
+				severity: 'error',
+			});
+		}
+	}
+
+	private async handleCustomerRoleAssignment(
+		operation: OperationNameEnum,
+		logToHelper: LogToHelper,
+		userLogtoId: string,
+		productName: string
+	) {
+		const roleAssignmentResponse = await logToHelper.assignCustomerPlanRoles(
+			userLogtoId,
+			productName.toLowerCase() as SupportedPlanTypes
+		);
+		if (roleAssignmentResponse.status !== 201) {
+			await eventTracker.notify({
+				message: EventTracker.compileBasicNotification(
+					`Failed to assign roles to user for planType ${productName}: ${roleAssignmentResponse.error}`,
+					operation
+				),
+				severity: 'error',
+			});
+			return;
+		}
+
+		await eventTracker.notify({
+			message: EventTracker.compileBasicNotification(
+				`${productName} plan assigned to user with logtoId ${userLogtoId}`,
+				operation
+			),
+			severity: 'info',
+		});
+	}
+
+	private async handleCustomerRoleRemoval(operation: OperationNameEnum, logto: LogToHelper, userLogtoId: string) {
+		const responses = await Promise.allSettled([
+			logto.removeLogtoRoleFromUser(userLogtoId, process.env.LOGTO_TESTNET_ROLE_ID.trim()),
+			logto.removeLogtoRoleFromUser(userLogtoId, process.env.LOGTO_MAINNET_ROLE_ID.trim()),
+		]);
+
+		const allRolesRemoved = responses.every((r) => r.status === 'fulfilled' && r.value.status === StatusCodes.OK);
+		if (allRolesRemoved) {
+			await eventTracker.notify({
+				message: EventTracker.compileBasicNotification(
+					`Roles have been removed successfully for user with id: ${userLogtoId}`,
+					operation
+				),
+				severity: 'info',
+			});
+			return;
+		}
+
+		for (const resp of responses) {
+			if (resp.status === 'rejected' || resp.value.status !== StatusCodes.OK) {
+				const errMsg = resp.status === 'rejected' ? (resp.reason as Error).message : resp.value.error;
+				await eventTracker.notify({
+					message: EventTracker.compileBasicNotification(`Role removal error: ${errMsg}`, operation),
+					severity: 'error',
+				});
+			}
+		}
+	}
+
+	private async syncLogtoRoles(operation: OperationNameEnum, customerId: string, productName: string) {
+		const logToHelper = new LogToHelper();
+		const setupResp = await logToHelper.setup();
+		if (setupResp.status !== StatusCodes.OK) {
+			await eventTracker.notify({
+				message: EventTracker.compileBasicNotification(
+					`Logto client initialisation failed: ${setupResp.error}`,
+					operation
+				),
+				severity: 'error',
+			});
+
+			return;
+		}
+
+		const user = await UserService.instance.userRepository.findOne({ where: { customer: { customerId } } });
+
+		if (user) {
+			switch (operation) {
+				case OperationNameEnum.SUBSCRIPTION_CREATE:
+				case OperationNameEnum.SUBSCRIPTION_UPDATE:
+					this.handleCustomerRoleAssignment(operation, logToHelper, user.logToId, productName);
+					return;
+				case OperationNameEnum.SUBSCRIPTION_CANCEL:
+					this.handleCustomerRoleRemoval(operation, logToHelper, user.logToId);
+					return;
+			}
+		}
+
+		await eventTracker.notify({
+			message: EventTracker.compileBasicNotification(
+				`Role assignment failed: No user found with customerId: ${customerId}`,
+				operation
+			),
+			severity: 'error',
+		});
 	}
 }
